+++ types/data_types.go
@@ -1,11 +1,8 @@
 package types
 
 import (
-	"time"
-
-	"github.com/datazip-inc/olake/constants"
-	"github.com/goccy/go-json"
 	"github.com/parquet-go/parquet-go"
+	"time"
 )
 
 type DataType string
@@ -94,178 +91,6 @@ type IcebergWriterPayload struct {
 	Metadata Metadata    `json:"metadata"`
 	Records  []IceRecord `json:"records,omitempty"` // omitempty because it's not in commit type
 }
-
-func (r *RawRecord) ToDebeziumFormat(db string, _ string, normalization bool, _ string) (*IceRecord, error) {
-	var rItem []IceColumn
-
-	// First create the schema and track field types
-	// schema := r.createDebeziumSchema(db, stream, normalization)
-
-	// Create the payload with the actual data
-	payload := make(map[string]interface{})
-
-	// Add olake_id to payload
-	payload[constants.OlakeID] = r.OlakeID
-
-	// Handle data based on normalization flag
-	if normalization {
-		for key, value := range r.Data {
-			if value == nil {
-				continue
-			}
-			rItem = append(rItem, IceColumn{
-				Key:     key,
-				Value:   value,
-				IceType: toIceServerType(value),
-			})
-		}
-	} else {
-		dataBytes, err := json.Marshal(r.Data)
-		if err != nil {
-			return nil, err
-		}
-		rItem = append(rItem, IceColumn{
-			Key:   "data",
-			Value: dataBytes,
-		})
-	}
-	r.OlakeTimestamp = time.Now().UTC()
-	rItem = append(rItem, IceColumn{Key: constants.OpType, Value: r.OperationType, IceType: toIceServerType(r.OperationType)})
-	rItem = append(rItem, IceColumn{Key: constants.DBName, Value: db, IceType: toIceServerType(db)})
-	rItem = append(rItem, IceColumn{Key: constants.CdcTimestamp, Value: r.CdcTimestamp.UTC(), IceType: toIceServerType(r.CdcTimestamp)})
-	rItem = append(rItem, IceColumn{Key: constants.OlakeTimestamp, Value: r.OlakeTimestamp, IceType: toIceServerType(r.OlakeTimestamp)})
-	rItem = append(rItem, IceColumn{Key: constants.OlakeID, Value: r.OlakeID, IceType: toIceServerType(r.OlakeID)})
-
-	return &IceRecord{
-		Record:     rItem,
-		RecordType: r.OperationType,
-	}, nil
-	// Add the metadata fields
-	// payload[constants.OpType] = r.OperationType // "r" for read/backfill, "c" for create, "u" for update
-	// payload[constants.DBName] = db
-	// payload[constants.CdcTimestamp] = r.CdcTimestamp
-	// payload[constants.OlakeTimestamp] = r.OlakeTimestamp
-
-	// Create Debezium format
-	// debeziumRecord := map[string]interface{}{
-	// 	"destination_table": stream,
-	// 	"key": map[string]interface{}{
-	// 		"schema": map[string]interface{}{
-	// 			"type": "struct",
-	// 			"fields": []map[string]interface{}{
-	// 				{
-	// 					"type":     "string",
-	// 					"optional": true,
-	// 					"field":    constants.OlakeID,
-	// 				},
-	// 			},
-	// 			"optional": false,
-	// 		},
-	// 		"payload": map[string]interface{}{
-	// 			constants.OlakeID: r.OlakeID,
-	// 		},
-	// 	},
-	// 	"value": map[string]interface{}{
-	// 		"schema":  schema,
-	// 		"payload": payload,
-	// 	},
-	// }
-
-	// // Add thread_id if not empty
-	// if threadID != "" {
-	// 	debeziumRecord["thread_id"] = threadID
-	// }
-}
-
-// func (r *RawRecord) createDebeziumSchema(db string, stream string, normalization bool) map[string]interface{} {
-// 	fields := make([]map[string]interface{}, 0)
-
-// 	// Add olake_id field first
-// 	fields = append(fields, map[string]interface{}{
-// 		"type":     "string",
-// 		"optional": true,
-// 		"field":    constants.OlakeID,
-// 	})
-
-// 	if normalization {
-// 		// Collect data fields for sorting
-// 		dataFields := make([]map[string]interface{}, 0, len(r.Data))
-
-// 		// Add individual data fields
-// 		for key, value := range r.Data {
-// 			field := map[string]interface{}{
-// 				"optional": true,
-// 				"field":    key,
-// 			}
-
-// 			switch value.(type) {
-// 			case bool:
-// 				field["type"] = "boolean"
-// 			case int, int8, int16, int32:
-// 				field["type"] = "int32"
-// 			case int64:
-// 				field["type"] = "int64"
-// 			case float32:
-// 				field["type"] = "float32"
-// 			case float64:
-// 				field["type"] = "float64"
-// 			case time.Time:
-// 				field["type"] = "timestamptz" // use with timezone as we use default utc
-// 			default:
-// 				field["type"] = "string"
-// 			}
-
-// 			dataFields = append(dataFields, field)
-// 		}
-
-// 		// Sorting basis on field names is needed because
-// 		// Iceberg writer detects different schemas for
-// 		// schema evolution based on order columns passed
-// 		sort.Slice(dataFields, func(i, j int) bool {
-// 			return dataFields[i]["field"].(string) < dataFields[j]["field"].(string)
-// 		})
-
-// 		fields = append(fields, dataFields...)
-// 	} else {
-// 		// For non-normalized mode, add a single data field as string
-// 		fields = append(fields, map[string]interface{}{
-// 			"type":     "string",
-// 			"optional": true,
-// 			"field":    "data",
-// 		})
-// 	}
-
-// 	// Add metadata fields
-// 	fields = append(fields, []map[string]interface{}{
-// 		{
-// 			"type":     "string",
-// 			"optional": true,
-// 			"field":    constants.OpType,
-// 		},
-// 		{
-// 			"type":     "string",
-// 			"optional": true,
-// 			"field":    constants.DBName,
-// 		},
-// 		{
-// 			"type":     "timestamptz",
-// 			"optional": true,
-// 			"field":    constants.CdcTimestamp,
-// 		},
-// 		{
-// 			"type":     "timestamptz",
-// 			"optional": true,
-// 			"field":    constants.OlakeTimestamp,
-// 		},
-// 	}...)
-
-// 	return map[string]interface{}{
-// 		"type":     "struct",
-// 		"fields":   fields,
-// 		"optional": false,
-// 		"name":     fmt.Sprintf("%s.%s", db, stream),
-// 	}
-// }
 
 func (d DataType) ToNewParquet() parquet.Node {
 	var n parquet.Node
